Feat: separating codes related to linked_list implementation from spinlock_module.c by adding new file named linked_list_impl.c
